move structures to separate headers
