+++ cache.h
@@ -14,40 +14,9 @@
 #define ALGORITHM_LINEAR 3
 #define ALGORITHM_NAIVE_OPTIMISTIC 4
 
-typedef struct elem
-{
-	struct elem *next;
-	struct elem *prev;
-	int set;
-	size_t delta;
-	char pad[32]; // up to 64B
-} Elem;
-
-struct
-config
-{
-	int rounds, cal_rounds;
-	int stride;
-	int cache_size;
-	int buffer_size;
-	int cache_way;
-	int cache_slices;
-	int threshold;
-	int algorithm;
-	int strategy;
-	int offset;
-	int con, noncon; // only for debug
-    void (*traverse)(Elem*);
-	double ratio;
-	int verbose, no_huge_pages, calibrate, retry, backtracking, ignoreslice, findallcolors, findallcongruent, verify, debug, conflictset;
-};
+#include "private_structs.h"
 
 #ifdef THREAD_COUNTER
-struct params_t {
-	pthread_mutex_t lock;
-	uint64_t counter;
-};
-
 static pthread_t thread;
 static struct params_t params;
